refactor: Restructure home feature UI and optimize analytics data flow
- Relocated dialogs and debug panels to dedicated `dialogs` and `debug` subpackages.
- Extracted `ProgressCard` and `TodayBlockingControls` from `HomeScreen.kt` into standalone components.
- Optimized analytics performance by fetching summarized daily totals instead of raw segments for long-range timeline data.
- Refined review prompt logic in `HomeViewModel` using modern reactive patterns.
- Updated `HomeScreen` pager and gesture logic to dynamically handle variable data ranges based on the user's first launch date.
- Centralized analytics constants and enums in `UsageAnalyticsModels.kt`.
